fix: over-zealous window-resize msg with clipboard
The window size alert is brought up (and kept up) if the terminal
size changes to something that can't accomodate the whole app. That
size included the clipboard message string size.

This means if you had the misfortune of resizing your terminal while
clipboard message was shown, you would need to expand the terminal to
the full message size for the "Terminal too small" to go away.